requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.2
async-timeout==4.0.2
billiard==4.1.0
black==23.3.0
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dill==0.3.6
dj-database-url==2.0.0
Django==4.1.7
django-ckeditor==6.5.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-stubs==1.16.0
django-stubs-ext==0.8.0
dotenv==0.0.5
et-xmlfile==1.1.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.6.0
isort==5.12.0
kombu==5.3.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
tornado==6.3.2
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
tzdata==2023.2
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
wrapt==1.15.0
